pyproject.toml
[tool.poetry]
name = "multipole_net"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poe.task]
force-torch = "python -m pip install torch==1.11.0+cu113 torchvision==0.12.0+cu113 torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu113"
force-scatter = "python -m pip install torch-scatter -f https://data.pyg.org/whl/torch-1.11.0+cu113.html"
force-sparse = "python -m pip install torch-sparse -f https://data.pyg.org/whl/torch-1.11.0+cu113.html"
force-geometric = "python -m pip install torch-geometric"

[tool.poetry.dependencies]
python = "=3.8.*"
tensorflow-gpu = "^2.4.0"
sacred = "^0.8.2"
seml = "^0.3.6"
scipy = "1.6.2"
matplotlib = "^3.5.1"
ipython = "^8.2.0"
jupyter = "^1.0.0"
pandas = "^1.4.2"
sympy = "^1.10.1"
nose = "^1.3.7"
tensorflow-addons = "^0.16.1"
ase = "^3.22.1"
lmdb = "^1.3.0"
loguru = "^0.6.0"
rdkit = "^2022.3.5"
networkx = "^2.8.5"
pydantic = "^1.9.2"
mayavi = "^4.8.0"
PyQt5 = "^5.15.7"
fire = "^0.4.0"
numba = "^0.56.2"
submitit = "^1.4.5"
wandb = "^0.13.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
